requirements.txt
python-decouple
discord